Model scripts renaming
